Updated layout and added validation
Removed table aspect and went with using divs.  I believe this will be easier when it comes to implementing the form and I can style into columns.  I also think this will be easier when dynamically adding and removing rows

Also added some light form validation as a place holder for examples of validation types/ways to validate