Added remaining input fields for round handicap
Added the remaining inputs for the round handicap

Made update to property names by removing underscore (_) and making everything camelCase

Adjusted added fields validation divs to display if the fields value is 0 (and to be hidden if the value is something OTHER than 0, i.e. a usable number).  This is a test/placeholder in case I want to adjust validation requirements in the future.  Decided I did not want to set a minimum number input as these inputs can be a wide range where as for round inputs you in theory could do all hole in 1's and have a score of 9 or 18 but could NEVER have a score less than that