Added simple calculation & output / updated strict initializiation
Added 1 global variable to be used to hold the final handicap value and set it to display on the page

Updated tsconfig file to remove error I was receiving when I was trying to add a global undefined variable.  Error was telling me to initialize in constructor but this is due to TS strict initialization.  Made change to remove that so I can add undefined global variables with no issue

Commented out the validation divs for the time being as not needed while still setting up