Update calculate handicap functionality
Btn to calculate handicap is disabled until the form is entirely valid.  This also addresses when a user does a calculation, then adds a row the btn becomes disabled because the form is 'invalid' due to the new row.  Once that new row is filled out and the entire form is valid again the btn becomes enabled