Enable/Disable Calc Handicap Btn
Added logic to enable and disable the calc handicap btn based on if the form is invalid, if the calculate handicap btn was clicked OR if there was a change made AFTER the handicap btn was clicked

created boolean calcBtnDisabled and if it's True then we disable the btn, if it's false we enable the btn

Set it to false by default and once the handicap btn is clicked we set it to true, therefore disabling the btn since we just created a handicap.  Then using the valueChanges we check to see if the calc Btn is disabled (i.e. if it's True) and if it is and there is a value change then we set calcBtnDisabled to false which ENABLES the btn so that we can re-calc the index as a change was made