Updated validation messages
Updated setting of the validation message so specific messages fire for specific input fields (i.e. message is different for 18 holes than 9 holes).

Previously it was looking for an control errors (like a minimum value not being met error) and it would look to pass the 'min' as a key into the object holding the key:value pairs of validation messages

Now the setValidationMessage method gets another parameter which tells the method if it is related to 9 or 18 hole input.  It then uses that parameter string as a key to find the corresponding key in the object for the specific validation message