Set simple minimum for values
Basic way to set minimum value for inputs.  Using 2 as placeholder but this will be updated to final values when finished

This will possibly be updated later.  Currently if you use the up/down arrows it starts at the min and won't go below it HOWEVER if you manually enter a value below the min you are allowed to enter that value (however you will get a validation message)