Added cross validation for formArray
used valueChanges observable to check if radio button value changes and if it does to then set the validator and pass in the NEW minimum value for the custom validator

2 ISSUES REMAIN:
In HTML I have 2 span tags that are identical except the required round value is different.  Want to only have 1 span tag and dynamically set that verbiage ==> Issue is when 2+ fields have validation messages and you change value of one form control and then of another (while messages are still there) the verbiage for the minimum score switches

User changes radio btn and enters in a score above that value and changes radio btn the validation doesn't re-fire.  Ex: user selects 9 holes, enters value of 9+, then changes radio btn to 18.  Validation doesn't re-fire and user COULD calculate handicap index.  If user makes change to value then the validation re-fires