Refactor dynamic validation for radio btns
Originally was doing logic checking on entire form group by looking for roundSelection changes.  Refactored so I am subscribing to any changes specifically on the roundSelection (opposed to the entire form).  Now if there is a change to roundSelection radio btn then it looks to see if the new 'value' (aka change) was to 9 or 18 holes and updates the validation message by passing in the new value for number of holes played

Misc. other refactoring by removing unused code