Reset values &  control errors on radio btn click/switch
Set user round score values back to null and remove all errors (which comes from custom validator) whenever a radio btn is clicked.  This will clear out the user input and force the user to re-enter a round score.  Clearing the errors also resets the validation message as well as it is dependent on an error being set from the custom validator