Fix bug relating to calc handicap btn
Calc handicap btn was enabled for click in scenario where:

all forms are valid and btn is clicked to calc handicap.  User changes radio btn from 18 to 9 and puts in a valid score (11 in this example) where it is above the current round but below the other round option.  User changes radio button and the calculate handicap btn is enabled even though the round value is null (it should be disabled in this scenario)

Added check to make sure that the calcBtnDisabled = false if the userRound score is NOT null AND to disable the btn when radio btn changes (which would mean the round score would become null and therefore no calculation can be done at that point in time)