Fix bug relating to radio btn's and validation msg
Removed setting errors to null on every change of radio btn.  This was allowing the form to stay valid even though on radio btn switch the user round score value was null.  This in part because on radio btn change we set the value to null.  Even though this value is null we are still setting a value which in turn keeps the form valid

By removing the ability to remove the error message, when user switches radio btns the value gets set to null and the validation message fires to enter in a minimum value

Was removing values before because validation was firing on change and now it fires on blur