Determine TRUE number of rounds
Created function that looks at all the round inputs selected and determines the TRUE number of rounds played.  One round of 18 is a true round and an even number of rounds of 9 is a true round (ex: 2 rounds of 9 counts as 1 true round | 3 rounds of 9 still only counts as 1 | 4 rounds of 9 count as 2 true rounds)

Took that total number of true rounds and stored in global variable that will be used to help calculate the TRUE handicap index

Commit also includes misc. fixes by re-naming some variables for clarification and moving the validation function up for better clarity