Update to resolve console warning
Console warning was firing on click of 'add round' btn:

It looks like you're constructing using a FormControl with both an options argument and an
  async validators argument. Mixing these arguments will cause your async validators to be dropped.
  You should either put all your validators in the options object, or in separate validators
  arguments

This error is in relation to the order of parameters for the form control.  I previously had the update action where the validators should have been

chatGPT suggestion was to create an object for the second param and pass in both validators and async operation (i.e. updateOn)