Refactor calculateDifferential method
Previously I was looping over the entire form array to get all the values, etc.  This was not efficient as this was triggering unnecessary calls/getting values of the form groups I was not using.

updated function to take in the formGroup index which is obtained when we build the form, we pass in an index position of that formGroup int he form Array

Now when we get values, look for changes, etc. we are only triggering it for that specific form group at the specified index position instead of looping through every since formGroup

Also updated the 'index' to 'formGroupIndex' to make it more descriptive and understand what 'index' realy represents

Added function to create a differential for 2 9 round scores.  THIS IS NOT DONE.  There is an issue with the array/logic that needs to be looked into further