Calculate 9 hole differentials
Updated function so now we add any 9 hole differentials to it's own array and place it in the array at it's formGroup index position.  This is so when we need to update the round score/differential we update it in the array at the appropriate position instead of just pushing/adding new differentials to the array

This will cause empty positions in the array so we use a filter method to return everything that isn't empty and then we sum the array up if there are 2 items in the array.  Then we wipe the array so it can be used again in the future