First cut at storing final differential
Add 9 hole differentials into an array just for 9 holes.  Loop through that array and sum up combination of consecutive values to create a finalized 18 hole differential

ISSUE TO RESOLVE: is storing the differentials in the CORRECT INDEX in the array.  Currently there is an issue when you do a 9 hole value, consecutive 18 hole values and then a second 9 hole value, that sum of the 9 hole values is overwritting an 18 hole differential