Fix broken calculation
Calculate Handicap function has a filter to combine the 18 and 9 hole differential arrays.  The issue was that it was not removing 0's from the array which was affecting which case statement to fire and incorrectly choosing the wrong number of differentials to do the calculation on

Additionally in the calculateCombinedDifferentials function we were combining 9 hole differentials and dividing by 2.  However per WHS we do NOT need to divide by 2 and just combine the differentials.  Removed the division